Omit ‘inline’
With today’s compilers ‘inline’ is typically not needed for
performance (at least the way GNU Tar uses it) and it gets in the
way of portability.  Some compilers ignore ‘inline’ and some
inline even without ‘inline’, and when inlining is haphazard GNU
Tar can link on your GNU/Linux host but not link on Solaris.
Better to have it not link everywhere so these problems are caught
before porting.
* configure.ac: Omit AC_C_INLINE; no longer needed here.
* lib/system.h (sys_reset_uid_gid): No longer inline.