Move sys_reset_uid_gid to library
This fixes a problem where ‘gcc -fanalyzer -flto’ (GCC 10)
complained that sys_reset_uid_gid was defined but not used in some
modules.  This function belonged in a .c file anyway.
* lib/rtapelib.c (sys_reset_uid_gid):
* paxlib/rtape.c (sys_reset_uid_gid):
Move here from system.h.  The code is now duplicated, but so is
most of the rest of this file anyway.
* lib/system.h (sys_reset_uid_gid): Remove.