More modernization of configuration
* am/rmt.m4 (PU_RMT): Do not bother checking for sgtty.h or
sys/socket.h.  Check for sys/mtio.h at most once.
* am/system.m4 (PU_SYSTEM): Do not check for headers like string.h
that we can now safely assume, or for headers like memory.h that
we no longer include, or for functions like lstat that Gnulib now
handles.  Check for headers just once.  Check for getaddrinfo
instead of the obsolete gethostbyname.
* bootstrap.conf (XGETTEXT_OPTIONS): Prefer Gnulib to gettext
for .m4 files.
* configure.ac: Omit checks that am/*.m4 or Gnulib now do for us,
or that are for obsolete platforms, or that we don’t need.  Bump
gettext version to 0.22.5.
* gnulib.modules: Add inttypes.
* lib/DISTFILES: Remove system-ioctl.h.
* lib/system-ioctl.h: Remove.  All uses simplifed to including
sys/ioctl.h and sys/mtio.h if HAVE_SYS_MTIO_H, since the other
parts of system-ioctl.h are no longer needed.
* lib/rmt.h (O_CREAT): Do not define; it’s safe to assume it now.
* lib/rtapelib.c (rmt_open__): Instead of using gethostbyname
which is obsolete, use getaddrinfo if available.
2024-07-25  Paul Eggert  <eggert@cs.ucla.edu>